Re #13540 Moved template functions to their own header
