+++ Code/Mantid/MantidQt/CustomInterfaces/inc/MantidQtCustomInterfaces/ReflGenerateNotebook.h
@@ -95,44 +95,6 @@ namespace MantidQt {
 
     std::tuple<std::string, std::string> transWSString(const std::string & trans_ws_str, const std::string & instrument);
 
-    /**
-    Create string of comma separated list of values from a vector
-    @param param_vec : vector of values
-    @return string of comma separated list of values
-    */
-    template<typename T, typename A>
-    std::string vectorString(const std::vector<T,A> &param_vec)
-    {
-      std::ostringstream vector_string;
-      const char* separator = "";
-      for(auto paramIt = param_vec.begin(); paramIt != param_vec.end(); ++paramIt)
-      {
-        vector_string << separator << *paramIt;
-        separator = ", ";
-      }
-
-      return vector_string.str();
-    }
-
-    /**
-    Create string of comma separated list of parameter values from a vector
-    @param param_name : name of the parameter we are creating a list of
-    @param param_vec : vector of parameter values
-    @return string of comma separated list of parameter values
-    */
-    template<typename T, typename A>
-    std::string vectorParamString(const std::string & param_name, const std::vector<T,A> &param_vec)
-    {
-      std::ostringstream param_vector_string;
-
-      param_vector_string << param_name << " = '";
-      param_vector_string << vectorString(param_vec);
-      param_vector_string << "'";
-
-      return param_vector_string.str();
-    }
-
-
     class DLLExport ReflGenerateNotebook {
 
     public:
